refactor: Extract RecipientList from WhoHasAccess
